move:#6 초기 세팅에 맞춰 파일 이동
